// src/app/models/car.ts
export interface Car {
  id: number;
  brand: number;
  brandName: string;
  brandImage: string;
  model: string;
  seller: number;
  sellerName?: string;
  sellerSurname?: string;
  sellerTelephone?: number;
  year: number;
  price?: number;
  fuel: number;
  fuelName?: string;
  doorNum: number;
  description?: string;
  image1?: string;
  image2?: string;
  image3?: string;
  date?: string;
  mileage?: number;
  engine?: string;
}